requirements.txt
streamlit
hugchat
streamlit-chat
streamlit-extras
pandas
numpy
nltk
spacy
textblob
emoji
bs4
contractions
pyspellchecker
openpyxl
transformers
langchain
langchain-community
chromadb
tiktoken
pypdf
sentence-transformers
pymupdf
python-dotenv
